split filter assy and wet at lhp and dashboard eff also rejection

+++ app/Views/pages/lhp_view.php
@@ -66,21 +66,22 @@
 												</tr>
 											</thead>
 											<tbody>
-												<?php foreach($data_lhp as $lhp) : ?>
-												<tr>
-													<!-- <td><?=$lhp['no_doc']?></td> -->
-													<td><?=$lhp['tanggal_produksi']?></td>
-													<td><?=$lhp['shift']?></td>
-													<td><?=($lhp['line'] == 10) ? 'MCB' : $lhp['line']?></td>
-													<td><?=$lhp['kasubsie']?></td>
-													<td><?=$lhp['nama_pic']?></td>
-													<!-- <td><?=$retVal = (!empty($lhp['total_aktual']) && !empty($lhp['total_plan'])) ? number_format((float) ($lhp['total_aktual'] / $lhp['total_plan']) * 100, 2, '.', '') : '' ; ?></td> -->
-													<td>
-														<a href="<?=base_url()?>lhp/detail_lhp/<?=$lhp['id_lhp_2']?>" class="btn btn-primary btn-sm" target="_blank">Detail</a>
-														<a href="<?=base_url()?>lhp/hapus_lhp/<?=$lhp['id_lhp_2']?>" class="btn btn-danger btn-sm" onclick="return confirm('Apakah anda yakin?')">Hapus</a>
-													</td>
-												</tr>
-												<?php endforeach; ?>
+												<?php foreach($data_lhp as $lhp) : 
+													if ($lhp['line'] != 8 && $lhp['line'] != 9) { ?>
+														<tr>
+															<!-- <td><?=$lhp['no_doc']?></td> -->
+															<td><?=$lhp['tanggal_produksi']?></td>
+															<td><?=$lhp['shift']?></td>
+															<td><?=($lhp['line'] == 10) ? 'MCB' : $lhp['line']?></td>
+															<td><?=$lhp['kasubsie']?></td>
+															<td><?=$lhp['nama_pic']?></td>
+															<!-- <td><?=$retVal = (!empty($lhp['total_aktual']) && !empty($lhp['total_plan'])) ? number_format((float) ($lhp['total_aktual'] / $lhp['total_plan']) * 100, 2, '.', '') : '' ; ?></td> -->
+															<td>
+																<a href="<?=base_url()?>lhp/detail_lhp/<?=$lhp['id_lhp_2']?>" class="btn btn-primary btn-sm" target="_blank">Detail</a>
+																<a href="<?=base_url()?>lhp/hapus_lhp/<?=$lhp['id_lhp_2']?>" class="btn btn-danger btn-sm" onclick="return confirm('Apakah anda yakin?')">Hapus</a>
+															</td>
+														</tr>
+												<?php } endforeach; ?>
 											</tbody>
 											<tfoot>
 												<tr>
